restructure

+++ util/utils.py
@@ -1,15 +1,18 @@
 import logging
-def mkfolder(folder):
+def mkfolder(folder, remove=True):
     import os
     try:
         os.makedirs(folder)
     except FileExistsError:
-        logging.warning("The {0} folder already exists  \n The old {0} folder will be renamed (to {0}~)".format(folder))
-        import shutil
-        if os.path.exists(folder+'~'):
-            shutil.rmtree(folder+'~')
-        os.system('mv {} {}'.format(folder, folder+'~'))
-        os.makedirs(folder)
+        if remove:
+            logging.warning(f"The {folder} folder already exists. The old {folder} folder will be moved to {folder}~")
+            import shutil
+            if os.path.exists(folder+'~'):
+                shutil.rmtree(folder+'~')
+            os.system('mv {} {}'.format(folder, folder+'~'))
+            os.makedirs(folder)
+        else:
+            logging.info(f"The {folder} folder already exists, outputs will write into this folder")
 
 
 def process_gpuID(gpuID):
